backend/requirements.txt
fastapi
uvicorn
supabase
pyarrow
pandas
python-multipart
python-jose[cryptography]
python-dotenv
xgboost
shap
sentence-transformers
pinecone
google-generativeai
